Evolucoes Microservices

+++ suporte/src/main/java/br/com/fiap/ecommerce/suporte/controller/ChamadoController.java
@@ -1,12 +1,23 @@
 package br.com.fiap.ecommerce.suporte.controller;
 
+import java.util.List;
+
+import javax.persistence.EntityNotFoundException;
+
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import br.com.fiap.ecommerce.suporte.model.Chamado;
 import br.com.fiap.ecommerce.suporte.service.ChamadoService;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("chamados")
@@ -18,28 +29,45 @@ public class ChamadoController {
 
 
     @GetMapping
-    public List<Chamado> findAll() {
-        return chamadoService.findAll();
+    public ResponseEntity<List<Chamado>> findAll() {
+        return ResponseEntity.ok(chamadoService.findAll());
     }
 
     @GetMapping("/{id}")
-    public Chamado findByCpf(@PathVariable("id") final Long id) {
-        return chamadoService.getPedido(id);
+    public ResponseEntity<Chamado> findById(@PathVariable("id") final Long id) {
+    	try {
+			return ResponseEntity.ok(chamadoService.getChamado(id));
+		} catch (EntityNotFoundException e) {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+		}
     }
 
     @PostMapping
-    public Chamado create(@RequestBody final Chamado chamado) {
-        return chamadoService.create(chamado);
+    public ResponseEntity<Chamado> create(@RequestBody final Chamado chamado) {
+    	return ResponseEntity.status(HttpStatus.CREATED).body(chamadoService.create(chamado));
     }
 
     @PutMapping("/{id}")
-    public Chamado update(@PathVariable("id") final Long id, @RequestBody final Chamado chamado) {
+    public ResponseEntity<Chamado> update(@PathVariable("id") final Long id, @RequestBody final Chamado chamado) {
     	chamado.setId(id);
-        return chamadoService.update(chamado);
+    	Chamado prod = chamadoService.getChamado(id);
+
+		if (prod == null) {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+		}
+
+		return ResponseEntity.ok(chamadoService.update(chamado));
     }
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable("id") final Long id) {
-    	chamadoService.delete(id);
+    public ResponseEntity<String> delete(@PathVariable("id") final Long id) {
+    	Chamado prod = chamadoService.getChamado(id);
+
+		if (prod == null) {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+		}
+
+		chamadoService.delete(id);
+		return ResponseEntity.ok(null);
     }
 }
